app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=app/requirements.txt pyproject.toml
#
aiohttp==3.8.6
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
alembic==1.12.0
    # via
    #   chuut (pyproject.toml)
    #   flask-migrate
aniso8601==9.0.1
    # via flask-restx
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==4.0.1
    # via
    #   chuut (pyproject.toml)
    #   paramiko
blinker==1.6.3
    # via flask
certifi==2023.7.22
    # via
    #   cloudinary
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via flask
cloudinary==1.36.0
    # via chuut (pyproject.toml)
cryptography==41.0.4
    # via paramiko
flask==3.0.0
    # via
    #   chuut (pyproject.toml)
    #   flask-cors
    #   flask-graphql
    #   flask-jwt-extended
    #   flask-migrate
    #   flask-restx
    #   flask-sqlalchemy
flask-blueprint==1.3.0
    # via chuut (pyproject.toml)
flask-cors==4.0.0
    # via chuut (pyproject.toml)
flask-graphql==2.0.1
    # via chuut (pyproject.toml)
flask-jwt-extended==4.5.3
    # via chuut (pyproject.toml)
flask-migrate==4.0.5
    # via chuut (pyproject.toml)
flask-restx==1.1.0
    # via chuut (pyproject.toml)
flask-sqlalchemy==3.1.1
    # via
    #   chuut (pyproject.toml)
    #   flask-migrate
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
graphql-core==2.3.2
    # via
    #   flask-graphql
    #   graphql-server-core
graphql-server-core==1.2.0
    # via flask-graphql
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jsonschema==4.19.1
    # via flask-restx
jsonschema-specifications==2023.7.1
    # via jsonschema
load-dotenv==0.1.0
    # via chuut (pyproject.toml)
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.20.1
    # via
    #   chuut (pyproject.toml)
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.29.0
    # via chuut (pyproject.toml)
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.2
    # via
    #   marshmallow
    #   marshmallow-sqlalchemy
    #   pytest
paramiko==3.3.1
    # via chuut (pyproject.toml)
pillow==10.0.1
    # via chuut (pyproject.toml)
pluggy==1.3.0
    # via pytest
promise==2.3
    # via
    #   graphql-core
    #   graphql-server-core
psycopg2==2.9.9
    # via chuut (pyproject.toml)
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via
    #   flask-jwt-extended
    #   twilio
pynacl==1.5.0
    # via paramiko
pytest==7.4.2
    # via chuut (pyproject.toml)
python-dotenv==1.0.0
    # via load-dotenv
pytz==2023.3.post1
    # via flask-restx
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   chuut (pyproject.toml)
    #   twilio
rpds-py==0.10.6
    # via
    #   jsonschema
    #   referencing
rx==1.6.3
    # via graphql-core
six==1.16.0
    # via
    #   cloudinary
    #   graphql-core
    #   promise
sqlalchemy==2.0.22
    # via
    #   alembic
    #   chuut (pyproject.toml)
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
twilio==8.9.1
    # via chuut (pyproject.toml)
typing-extensions==4.8.0
    # via
    #   alembic
    #   sqlalchemy
urllib3==2.0.6
    # via
    #   cloudinary
    #   requests
uwsgi==2.0.22
    # via chuut (pyproject.toml)
werkzeug==3.0.0
    # via
    #   flask
    #   flask-jwt-extended
    #   flask-restx
yarl==1.9.2
    # via aiohttp
